+++ src/app/locales/en/components/qos-evaluation-info.js
@@ -1,5 +1,5 @@
 export default {
   invalidQosExpression: 'Invalid QoS expression',
-  qosEvaluationUknownError: 'Unknown QoS validation error',
+  qosEvaluationUnknownError: 'Unknown QoS validation error',
   invalidUnknown: 'unknown error',
 };