+++ src/app/services/production/file-manager.js
@@ -24,9 +24,26 @@ import createThrottledFunction from 'onedata-gui-common/utils/create-throttled-f
  * @param {Array<string>} files
  */
 
+/**
+ * @typedef {Object} RecallLogData
+ * @property {string} fileId CDMI Object ID of file that message is about
+ * @property {string} relativePath relative path to error-affected file from archive
+ * @property {RecallError} reason object with error reason
+ */
+
+/**
+ * An error object from backend in standard format, parsable using ErrorExtractor
+ * @typedef {RecallError} Object
+ */
+
+/**
+ * @typedef {JsonInfiniteLogEntry<RecallLogData>} RecallLogEntry
+ */
+
 const cancelRecallAspect = 'cancel_archive_recall';
 const childrenAttrsAspect = 'children_details';
 const symlinkTargetAttrsAspect = 'symlink_target';
+const recallLogAspect = 'archive_recall_log';
 const fileModelName = 'file';
 
 export default Service.extend({
@@ -34,6 +51,7 @@ export default Service.extend({
   onedataRpc: service(),
   onedataGraph: service(),
   timeSeriesManager: service(),
+  infiniteLogManager: service(),
   apiSamplesManager: service(),
 
   /**
@@ -491,6 +509,23 @@ export default Service.extend({
       gri: requestGri,
       subscribe: false,
     });
+  },
+
+  /**
+   * Begins a procedure of cancelling archive recall process that has root in
+   * file with `recallRootId` entity ID.
+   * @param {string} recallRootId
+   * @param {JsonInfiniteLogPagingParams} pagingParams
+   * @returns {Promise<JsonInfiniteLogPage<RecallLogData>>}
+   */
+  async getRecallLogs(recallRootId, pagingParams) {
+    const infiniteLogManager = this.get('infiniteLogManager');
+    const requestGri = gri({
+      entityType: fileEntityType,
+      entityId: recallRootId,
+      aspect: recallLogAspect,
+    });
+    return await infiniteLogManager.getJsonInfiniteLogContent(requestGri, pagingParams);
   },
 
   // TODO: VFS-7643 move browser non-file-model-specific methods to other service